Added API as an option to the installation, reorganized the API and fused all documentations in README.md
