+++ setup.py
@@ -52,6 +52,16 @@ def read_requirements():
     return deps
 
 
+def read_api_requirements():
+    api_deps = []
+    try:
+        with open("./lightrag/api/requirements.txt") as f:
+            api_deps = [line.strip() for line in f if line.strip()]
+    except FileNotFoundError:
+        print("Warning: API requirements.txt not found.")
+    return api_deps
+
+
 metadata = retrieve_metadata()
 long_description = read_long_description()
 requirements = read_requirements()
@@ -84,5 +94,15 @@ setuptools.setup(
         "Tracker": f"{metadata.get('__url__', '')}/issues"
         if metadata.get("__url__")
         else "",
+    },
+    extras_require={
+        "api": read_api_requirements(),  # API requirements as optional
+    },
+    entry_points={
+        "console_scripts": [
+            "lollms-lightrag-server=lightrag.api.lollms_lightrag_server:main [api]",
+            "ollama-lightrag-server=lightrag.api.ollama_lightrag_server:main [api]",
+            "openai-lightrag-server=lightrag.api.openai_lightrag_server:main [api]",
+        ],
     },
 )
